Update add-subscription.md
The variable `recipe` is supposed to point at a single `Recipe` instance or `None`, if no matching recipe was found. Currently, the `receipe` variable points to a list of `Recipe`s although the variable name does not reflect this. This PR addresses this flaw by using the built-in `next()` function.